<feat>(main, game): Crop images to expedite finding cards.
A problem that will arise is finding unknown cards when moving them from the stock to the waste. To do this, each card must individually be found against the entire screen, which would take too long. The solution is to crop the image to each area of the screen, i.e. the waste.

Additionally, an offset must be added to compensate for the cropped image positions.