<feat>(game, strategy): Add check to avoid duplicate game states.
A game state may be seen more than once, such as in the case of a red queen bouncing between two black kings. A check is implemented to undo if a game state has already been seen. Additionally, the check ensures that the last move was not an undo to avoid getting stuck in a branch.