<feat>(game): Optimize finding cards in stock.
Before the change, the stock is reset every time a card is drawn from it. After the change, the stock will not reset. Instead, to draw a card from the stock, the following the top waste card and the stock are checked. If the card is not in either places, the stock will finally reset and run through the stock.