<feat>(adb): Add point() method to assist with debugging positions. Adjust positions and timing for taps and swipes.
It is useful to display an image with a point where the program may think there is a card. The point() method creates a small blue dot at a given set of coordinates. Taps and swipes are given extra delay after the movement to allow the cards to animate properly. Swiped cards do not need as much leniency, so it can be reduced.