<feat, fix>(game): Add __repr__ to print game state instead of using print_state() method. Adjust faceup card offset when calculating card positions in tableau. Update card positions when moving to foundation. Use foundation position instead of card position when moving from foundation.
Faceup card offset was calculated incorrectly. The reference image had all 13 cards in a tableau column, which squished the cards together because there was not enough room. In practice, this position would never occur, so the offset was corrected. When cards are moved to the foundation, the card positions were not properly updated. In addition, when moving cards to the foundation, the constant foundation position was not used. Fixing both of these issues will create consistency when moving cards to and from the foundation.