<feat>(strategy): Create list to keep moves in order of priority. Print this list before choosing a move.
Printing a list of all moves in order of priority allows easier debugging of undos. The states log also keeps track of the available moves list at the time. All functionality to pick the moves are moved to a method order_moves(). Meanwhile, play_move() is rewritten to simply pick the first move in the ordered list.